requirements.txt
# Flask web framework
Flask==3.0.0
flask-cors==4.0.0

# AI and ML
google-generativeai==0.3.1

# Vector database
qdrant-client==1.7.0

# Data processing
pandas==2.1.0
numpy==1.24.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Additional utilities
python-dateutil==2.8.2
scikit-learn==1.3.0
